+++ src/main/java/dev/tomerdev/mercfrontcore/AddonConstants.java
@@ -7,7 +7,7 @@ import com.mojang.logging.LogUtils;
 public final class AddonConstants {
     public static final String MOD_ID = "mercfrontcore";
     public static final String MOD_NAME = "MERCFront-core";
-    public static final String MOD_VERSION = "1.0.5-0.7.1.2b";
+    public static final String MOD_VERSION = "1.0.6-0.7.1.2b";
     public static final Logger LOGGER = LogUtils.getLogger();
 
     private AddonConstants() {